hardcode hashsize and tracking num of exported/imported
hash table size doesn't represent # of entries (imported/exported).
It defines # of buckets. And each bucket is head of the linked list
of entries as many as needed.

So prevent confusion, we remove constants for it and just hardcode
the size.

Also, added global counter for tracking # of exported and imported
and make those displayed whenever added/removed

Signed-off-by: Dongwon Kim <dongwon.kim@intel.com>